All by the moment is Right

+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: jsaldana <jsaldana@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/31 11:14:58 by jsaldana          #+#    #+#             */
-/*   Updated: 2022/09/11 16:51:45 by jsaldana         ###   ########.fr       */
+/*   Updated: 2022/09/18 11:22:10 by jsaldana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,14 +14,14 @@
 
 int	ft_strncmp(const char *str1, const char *str2, size_t n)
 {
-	int				count;
+	size_t			count;
 	unsigned char	*s1;
 	unsigned char	*s2;
 
 	s1 = (unsigned char *)str1;
 	s2 = (unsigned char *)str2;
 	count = 0;
-	while ((str1[count] != '\0' || str2[count] != '\0') && count < (int)n)
+	while ((str1[count] != '\0' || str2[count] != '\0') && count < n)
 	{
 		if (s1[count] != s2[count])
 		{
